+++ CHANGELOG.md
@@ -15,6 +15,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added an `include_infinite` option to `FloatSpecs` (behind `fuzzing` feature gate)
 - Added length specifications to `StringSpecs` (behind `fuzzing` feature gate)
 - Added more public methods to `ByteSpecs` (behind `fuzzing` feature gate)
+- Added additional fields for customizing the operation in `CigletSpecs` (behind `fuzzing` feature gate)
 
 ### Changed